Remove lml_namespace dependency
Copy the namespace.hpp with lml default in tree here instead of
using a small dependency.

I'd deleted the lml_namespace repo during a github repo cleanout,
forgetting it was used by parameta. The namepace repo is a nice
idea but it's simpler to copy and paste the file where needed.